+++ src/Components/Data.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import USData from './DataComponent/USData'
+import CountryData from './DataComponent/CountryData'
 import StateData from './DataComponent/StateData'
 import { Container } from 'react-bootstrap'
 import Loading from './Loading'
@@ -8,7 +8,7 @@ const Data = (props) => {
 
   return (
     <Container  className='Data'>
-      <USData />
+      <CountryData />
       <StateData />
     </Container>
   )
