--- src/styles/components-scss/dataComponent-scss/_usData.scss
@@ -1,8 +1,0 @@
-@import '../../base-scss/mixins';
-@import '../../base-scss/variables';
-
-.USData {
-  border: 1px black solid;
-  height: $topSectionHeight;
-  margin-bottom: $sectionMargin;
-}
