+++ src/styles/main.scss
@@ -9,7 +9,7 @@
 
 @import '../styles/components-scss/data';
 @import '../styles/components-scss/dataComponent-scss/stateData';
-@import '../styles/components-scss/dataComponent-scss/usData';
+@import '../styles/components-scss/dataComponent-scss/countryData';
 
 @import '../styles/components-scss/media';
 @import '../styles/components-scss/mediaComponent-scss/articles';
